spi-nor: Workaround freescale hang on BMC
Attempts to read the SFDP (added post v4.13) hang the system with an AHB
timeout. The Freescale SPI driver is deprecated post v4.20, and so rather
than digging into this, simply use the SPI_NOR_SKIP_SFDP flag since we
are a fixed configuration.

This is obviously not suitable for unstreaming, but there are similar patches
for some quad-mode commands. Fix this in the new SPI driver when we migrate
past V4.20.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>