platina-example-amd64_config: Enable CONFIG_MODULES
In bac977cbc, the creation of the kernel headers package was made
conditoinal on the setting of CONFIG_MODULES, to fix a related build
issue.

Our build system assumes that there will be linux_headers_ for each
platform. This should be fixed when we move to using the standard
kernel configurations from Debian, but for now enable (but don't
enable any modules) CONFIG_MODULES to work around this.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>